feat 1.0.0 (feature-despacho): Agregando módulo Despacho y correcciones - Modificación de tipos de ID en la base de datos de int a bigint. - Cambio de tipo de dato del ID en el modelo de Conductor y ajustes correspondientes en DTO y Controller. - Ajuste del tipo de letra a camelCase en el módulo de Movimiento. - Cambio del nombre del archivo properties de application-TyronePilayMenendez.properties a application-devTyrone.properties. - Integración parcial de módulos Contenedores e InventarioContenedores para su uso en Despacho, implementando solo el método POST con su DTO y Controller. - Implementación del módulo Despacho con un CRUD básico y sólido.

+++ src/main/java/com/colaborativo/modulo_lyd/model/conductor/DatosListarConductor.java
@@ -6,10 +6,10 @@ package com.colaborativo.modulo_lyd.model.conductor;
  * Representa datos de un conductor utilizado para listar información.
  *
  * Cambios:
- * - Se cambió el tipo de datos de 'experiencia' para su compatibilidad en la db.
+ * - se cambio el tipo de dato de ID.
  */
 public record DatosListarConductor(
-        Integer idConductor,
+        Long idConductor,
         String nombreConductor,
         String licenciaVigente,
         String twicCard,
